perf(ui): speed up list view rendering, ensure root layout does not re-render when navigating to list view
